Added/Updated examples w.r.t 19.9

+++ Examples/SlidesCPP/GetCameraEffectiveData.cpp
@@ -33,7 +33,7 @@ void GetCameraEffectiveData()
 	System::SharedPtr<IThreeDFormatEffectiveData> threeDEffectiveData = pres->get_Slides()->idx_get(0)->get_Shapes()->idx_get(0)->get_ThreeDFormat()->GetEffective();
 
 	System::Console::WriteLine(u"= Effective camera properties =");
-	System::Console::WriteLine(System::String(u"Type: ") + System::ObjectExt::ToString(threeDEffectiveData->get_Camera()->get_CameraType()));
+	System::Console::WriteLine(System::String(u"Type: ") + System::ObjectExt::ToString((int)threeDEffectiveData->get_Camera()->get_CameraType()));
 	System::Console::WriteLine(System::String(u"Field of view: ") + threeDEffectiveData->get_Camera()->get_FieldOfViewAngle());
 	System::Console::WriteLine(System::String(u"Zoom: ") + threeDEffectiveData->get_Camera()->get_Zoom());
